[eas-cli] relocate getProjectIdAsync
