+++ packages/eas-cli/src/project/projectUtils.ts
@@ -2,6 +2,7 @@ import { getConfig } from '@expo/config';
 import pkgDir from 'pkg-dir';
 
 import { ensureLoggedInAsync } from '../user/actions';
+import { ensureProjectExistsAsync } from './ensureProjectExists';
 
 export async function getProjectAccountNameAsync(projectDir: string): Promise<string> {
   const { exp } = getConfig(projectDir, { skipSDKVersionRequirement: true });
@@ -12,4 +13,12 @@ export async function getProjectAccountNameAsync(projectDir: string): Promise<st
 export async function findProjectRootAsync(cwd?: string): Promise<string | null> {
   const projectRootDir = await pkgDir(cwd);
   return projectRootDir ?? null;
+}
+
+export async function getProjectIdAsync(projectDir: string): Promise<string> {
+  const { exp } = getConfig(projectDir, { skipSDKVersionRequirement: true });
+  return await ensureProjectExistsAsync({
+    accountName: await getProjectAccountNameAsync(projectDir),
+    projectName: exp.slug,
+  });
 }
